+++ examples/test2.php
@@ -1,34 +1,34 @@
 <?php
-
 include "../src/Playsms/Tpl.php";
 
 $the_data = 'This is the data';
 
 $tpl = new Playsms\Tpl;
 
-$tpl->name = 'dir2/test2';
-
-$tpl->vars = array(
+$tpl->setName('dir2/test2')->setVars(array(
 	'title' => 'This is test 2',
 	'content' => 'This is sample content',
-);
-
-$tpl->ifs = array(
+))->setIfs(array(
 	'valid' => TRUE,
 	'something' => FALSE,
-);
-
-$tpl->loops = array(
+))->setLoops(array(
 	'data' => array(
-		array('color' => 'Red',   'hex' => '#FF0000'),
-		array('color' => 'Green', 'hex' => '#00FF00'),
-		array('color' => 'Blue',  'hex' => '#0000FF'),
-	),
-);
-
-$tpl->injects = array('the_data');
-
-$tpl->compile();
+		array(
+			'color' => 'Red',
+			'hex' => '#FF0000'
+		) ,
+		array(
+			'color' => 'Green',
+			'hex' => '#00FF00'
+		) ,
+		array(
+			'color' => 'Blue',
+			'hex' => '#0000FF'
+		) ,
+	) ,
+))->setInjects(array(
+	'the_data'
+))->compile();
 
 echo "<p>Original content:</p>\n";
 echo $tpl->getContent();